refactor(next/dev): reusable start_server
